RMLNQSQL-64 Implemented recursive resolution of joined tables.
Currently, this code is not exercised by the actual system because the only place where a SqlJoin is added to a SqlTable is in DefaultMappingResolutionStage.ResolveEntityRefMemberExpression, and that method immediately resolves the new joined table. However, I still added it for completeness and correctness if some code later decides that it should add joined tables in other places.